Refactor adresse entity subtypes
