+++ lib/asp/entities/prestadoss.rb
@@ -31,7 +31,7 @@ module ASP
       def fragment(xml)
         prestadoss_xml(xml)
 
-        xml.adressesprestadoss { Adresse.from_payment_request(payment_request).to_xml(xml) }
+        xml.adressesprestadoss { adresse_entity_class.from_payment_request(payment_request).to_xml(xml) }
         xml.coordpaiesprestadoss { CoordPaie.from_payment_request(payment_request).to_xml(xml) }
         xml.listeelementpaiement { ElementPaiement.from_payment_request(payment_request).to_xml(xml) }
       end
