+++ spec/lib/asp/entities/enregistrement_spec.rb
@@ -6,7 +6,7 @@ describe ASP::Entities::Enregistrement, type: :model do
   let(:payment_request) { create(:asp_payment_request, :ready) }
 
   before do
-    %w[PersPhysique Adresse CoordPaie Dossier].each { |name| mock_entity(name) }
+    %w[PersPhysique Adresse::Base CoordPaie Dossier].each { |name| mock_entity(name) }
   end
 
   it_behaves_like "an ASP payment mapping entity"
