+++ spec/services/asp/mappers/adresse/base_mapper_spec.rb
@@ -2,7 +2,7 @@
 
 require "rails_helper"
 
-describe ASP::Mappers::AdresseMapper do
+describe ASP::Mappers::Adresse::BaseMapper do
   subject(:mapper) { described_class.new(payment_request) }
 
   let(:payment_request) { create(:asp_payment_request) }
